rename core/results.py into core/summary.py
