+++ DRE/core/__init__.py
@@ -1,7 +1,7 @@
 from .models import ModelsCube
 from .dre_cpu import ModelCPU, Parallelize
 from .cuts import Cutter
-from .results import Summary
+from .summary import Summary
 from . import statistics
 
 __all__ = ['ModelsCube', 'ModelCPU', 'Parallelize', 'Cutter', 'Summary', 'statistics']
